Split union.test.js into unionGeom2.test.js and unionGeom3.test.js
